Add orScream function for decoration failures
This function is a smaller combinator for decorating failures by
prepending them with a message.  This is helpful when you need to
provide more context on a specific expectation with an `it` block. It can
also be used to dissambiguate failures when looking with an expectation.

The originally implementation was:

orScream :: MonadCatch m => m a -> String -> m a
orScream action message =
  action `catch` \(HUnitFailure l r) ->
    throwM . HUnitFailure l . Reason $ message ++ "\n\n" ++ formatFailureReason r